requirements.txt
langchain
openai
streamlit
faiss-cpu
python-dotenv
